Conclusion
In this exercise we try to guess which input is the most important to focus our ads targeting. We first target on people features.
We make our machine learning training.
At the end we make diagram and we figure out which feature is important and what type of person is more apt to subscribe (YES in our excel datasheet).

We then extract the best feature of selection and check the result.

After evaluation we can see that we might go targeting more the person who has more than 75% of probability to subscribe at the end.
We can also see that the person under 24 that aren't student are also a good target with 25% of subscribing.