Conclusion
First we analyze we all the components : We see that Monday isn't a great day to do buisness as April.

If we analyse it without month we can observe that Monday isn't a good day but Wednesday is better.
If we analyse it without day we observe that march is a good month to make buisness(+33%).

By making experiment with various plot and playing with number of tree we can observe some things.
Without both month and day we observe that the success of a previous ads has more chance to get a subscribe as other (+50%).
We can also see that the index of confidence is accurate as it improve the chance of getting subscriber. So it is a tool to use more often.